+++ +sbfsem/+ui/LayoutManager.m
@@ -2,22 +2,22 @@ classdef LayoutManager < handle
 	
 	methods (Static)
 
-		function h = verticalBoxWithLabel(parentHandle, str, varargin)
+		function [h, p] = verticalBoxWithLabel(parentHandle, str, varargin)
 			% VERTICALBOXWITHLABEL
 
-			h = uix.VBox('Parent', parentHandle,...
+			p = uix.VBox('Parent', parentHandle,...
 				'BackgroundColor', 'w');
-			uicontrol(h, 'Style', 'text', 'String', str);
-			uicontrol(h, varargin{:});
-			set(h, 'Heights', [-0.75, -1]);
+			h = uicontrol(p, 'Style', 'text', 'String', str);
+			uicontrol(p, varargin{:});
+			set(p, 'Heights', [-0.75, -1]);
         end
         
-        function h = horizontalBoxWithLabel(parentHandle, str, varargin)
-            h = uix.HBox('Parent', parentHandle,...
+        function [h, p] = horizontalBoxWithLabel(parentHandle, str, varargin)
+            p = uix.HBox('Parent', parentHandle,...
                 'BackgroundColor', 'w');
-            uicontrol(h, 'Style', 'text', 'String', str);
-            uicontrol(h, varargin{:});
-            set(h, 'Widths', [-1, -1]);
+            h = uicontrol(p, 'Style', 'text', 'String', str);
+            uicontrol(p, varargin{:});
+            set(p, 'Widths', [-1, -1]);
         end
 	end
 end